// apps/server/src/male/types/maleTypes.ts
export type MaleType = {
    firstname: string;
    lastname: string;
    birthdate: Date;
    maritalstatus: string;
    fathername: string;
    mothername: string;
    origin: string;
    height: number;
    headwear: string;
    pelkoshers: string;
    currentaddress: string;
    imglink?: string;
    torahstudystatus: string;
    yeshiva: string;
    highereducation: string;
    educationname: string;
    highereducationacademy: string;
    jobstatus: string;
    jobcompany?: string;
    phonenumber: string;
    email: string;
    password: string;
    gender: string
  };
  